:fire: [#2555] -- remove dead UI templatetags code
