--- src/openforms/ui/templates/ui/components/anchor/anchor.html
@@ -1,8 +1,0 @@
-{% load ui %}{% spaceless %}
-    <a
-        class="utrecht-link utrecht-link--openforms{% if style %} utrecht-link--openforms-{{ style }}{% endif %}{% if active %} utrecht-link--openforms-active{% endif %}{% if hover %} utrecht-link--openforms-hover{% endif %}"
-        href="{{ href }}"
-        target="{{ target }}">{% spaceless %}
-        {{ label }}
-    {% endspaceless %}</a>
-{% endspaceless %}
