+++ src/openforms/ui/templates/ui/views/abstract/list.html
@@ -16,7 +16,10 @@
                         <ul class="list">
                             {% for object in object_list %}
                                 <li class="list__item">
-                                    {% anchor href=object.get_absolute_url label=object %}
+                                    <a
+                                        href="{{ object.get_absolute_url }}"
+                                        class="utrecht-link utrecht-link--openforms"
+                                    >{{ object }}</a>
                                 </li>
                             {% endfor %}
                         </ul>
@@ -25,6 +28,6 @@
             {% endblock %}
         </div>
 
-        {% a11y_toolbar user=user %}
+        {% a11y_toolbar %}
     </section>
 {% endblock %}
